requirements.txt
# Football Prediction Reels Template - Requirements
# Install with: pip3 install -r requirements.txt

# Image processing
Pillow>=9.0.0

# Optional: For advanced features
# numpy>=1.21.0
# opencv-python>=4.5.0
